feat(ansible): add filter for escaping double quotes for Terraform
